backend/requirements.txt
flask==2.0.1
torch==2.0.1
torchvision==0.15.2
ftfy==6.1.1
regex==2023.5.5
tqdm==4.65.0
pillow==9.5.0
transformers==4.30.2
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
datasets==2.12.0 
